Add bdd tests for goods ratings endpoint
