+++ api/data_workspace/v2/tests/bdd/test_goods_ratings.py
@@ -1,0 +1,4 @@
+from pytest_bdd import scenarios
+
+
+scenarios("./scenarios/goods_ratings.feature")
